short used

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,8 +6,8 @@
  */
 int main(void)
 {
-	int sum;
-	int first, second;
+	short int sum;
+	short int first, second;
 	int count;
 
 	sum = 0;
